Appi
Se creó la relación de la tabla de carga y rancho.

+++ Project/src/main/java/mx/LemonTrees/Project/Repository/CargaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import mx.LemonTrees.Project.Model.Carga;
 
 @Repository
-public interface CargaRepository extends CrudRepository<Carga,Integer>{
+public interface CargaRepository extends CrudRepository<Carga,Long>{
 
 }
